crypto-scalper/requirements.txt
ccxt==4.2.25
pandas==2.1.4
numpy==1.25.2
pyyaml==6.0.1
python-dotenv==1.0.0
ta==0.10.2
requests==2.31.0
websocket-client==1.6.4